Created generic API Client to work with all files
